# By jeor0980 (24) and others
# Via GitHub (11) and others
* dev: (53 commits)
  Update changelog and increment version.
  Add categories that where not in initial list. Add size for homepage
  Updating CHANGELOG and bumping version number in config.py
  Fixing spelling error in ds field for pantheon_size
  Updating atlas/version to release/v2.3.0-beta6
  Updating changelog for v2.3.0-beta6 release
  Add fields for site type and pantheon size. Resolves #601.
  Remove functions to add packages into default profiles.
  Update mysql connector
  Run a registry rebuild if we are removing all code to help with bundle clean up.
  Updating mysql-connector to version 2.1.7
  Remove whitespace.
  Restoring
  Removing deploy_php_cache_clear variable from tasks.py
  Updating CHANGELOG
  removing instances of clear_php_cache
  Update version to next one
  Updating changelog for v2.3.0-beta3
  Bring settings dict in as well.
  Coding standards and removing some debugging logs.
  ...

# Conflicts:
#	requirements.txt

+++ atlas/code_operations.py
@@ -10,7 +10,7 @@ import subprocess
 import git
 
 from atlas import utilities
-from atlas.config import (ENVIRONMENT, CODE_ROOT, WEB_ROOT)
+from atlas.config import (ENVIRONMENT, CODE_ROOT, WEB_ROOT, DEFAULT_PROFILE)
 from atlas.config_servers import (SERVERDEFS)
 
 # Setup a sub-logger. See tasks.py for longer comment.
@@ -66,17 +66,33 @@ def repository_remove(item):
 
 def update_symlink_current(item):
     """
-    Determine the path for a code item
+    Create symlink between version number directory and current
     """
+    # Build path for -current symlink
     code_folder_current = '{0}/{1}/{2}/{2}-current'.format(
         CODE_ROOT,
         utilities.code_type_directory_name(item['meta']['code_type']),
         item['meta']['name'])
-    # Remove symlink if it exists
-    if os.path.islink(code_folder_current):
-        os.unlink(code_folder_current)
-    os.symlink(utilities.code_path(item), code_folder_current)
-    log.debug('Code deploy | Symlink | %s', code_folder_current)
+
+    # Only link item if it is current
+    if item['meta']['is_current']:
+        # Remove previous, symlink if it exists
+        if os.path.islink(code_folder_current):
+            os.unlink(code_folder_current)
+        # Create symlink
+        os.symlink(utilities.code_path(item), code_folder_current)
+        log.debug('Code deploy | Symlink | %s', code_folder_current)
+
+    # If updated item to is_current:false don't remove the code-current directory of the same code
+    # item.
+    else:
+        package_query = 'where={{"meta.name":"{0}","meta.is_current":true}}'.format(
+            item['meta']['name'])
+        current_package_items = utilities.get_eve('code', package_query)
+
+        if not item['meta']['is_current'] and not current_package_items['_meta']['total']:
+            if os.path.islink(code_folder_current):
+                os.unlink(code_folder_current)
 
 
 def sync_code():
@@ -92,7 +108,7 @@ def sync_code():
 
 def deploy_static(item):
     """Deploy static asset to the web root
-    
+
     Arguments:
         item {dict} -- Item record for static asset to deploy
     """
